Move Java unit tests to DafnyRuntimeJava
