+++ Source/DafnyRuntime/DafnyRuntimeJava/pom.xml
@@ -14,6 +14,14 @@
         <maven.compiler.target>1.8</maven.compiler.target>
     </properties>
 
+    <dependencies>
+        <dependency>
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
+            <version>4.12</version>
+        </dependency>
+    </dependencies>
+
     <build>
   <plugins>
     <plugin>
