+++ Source/DafnyRuntime/DafnyRuntimeJava/src/test/java/BigOrdinalTest.java
@@ -1,4 +1,4 @@
-import DafnyClasses.BigOrdinal;
+import dafny.*;
 import org.junit.Test;
 
 import java.math.BigInteger;
