+++ Source/DafnyRuntime/DafnyRuntimeJava/src/test/java/ByteTest.java
@@ -1,4 +1,4 @@
-import DafnyClasses.UByte;
+import dafny.UByte;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
