+++ Source/DafnyRuntime/DafnyRuntimeJava/src/test/java/ULongTest.java
@@ -1,4 +1,4 @@
-import DafnyClasses.ULong;
+import dafny.ULong;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
